refactor: util 클래스 패키지 이동 ( #68 )
- util 패키지를 생성하여, 기존에 존재하던 util 클래스는 해당 패키지로 이동
