+++ src/main/java/jpabook/dashdine/security/filter/JwtAuthenticationFilter.java
@@ -8,9 +8,9 @@ import jakarta.servlet.http.HttpServletResponse;
 import jpabook.dashdine.domain.user.UserRoleEnum;
 import jpabook.dashdine.dto.request.user.LoginRequestDto;
 import jpabook.dashdine.dto.response.ApiResponseDto;
-import jpabook.dashdine.jwt.JwtUtil;
-import jpabook.dashdine.redis.RedisUtil;
+import jpabook.dashdine.util.JwtUtil;
 import jpabook.dashdine.security.userdetails.UserDetailsImpl;
+import jpabook.dashdine.util.RedisUtil;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
