+++ src/main/java/jpabook/dashdine/security/filter/JwtAuthorizationFilter.java
@@ -7,10 +7,10 @@ import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import jpabook.dashdine.domain.user.User;
-import jpabook.dashdine.jwt.JwtUtil;
-import jpabook.dashdine.redis.RedisUtil;
+import jpabook.dashdine.util.JwtUtil;
 import jpabook.dashdine.security.userdetails.UserDetailsServiceImpl;
 import jpabook.dashdine.service.user.query.UserQueryService;
+import jpabook.dashdine.util.RedisUtil;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
