+++ src/main/java/jpabook/dashdine/service/email/EmailManagementService.java
@@ -4,8 +4,8 @@ import jpabook.dashdine.domain.user.User;
 import jpabook.dashdine.dto.request.user.AuthCodeVerificationRequestDto;
 import jpabook.dashdine.dto.request.user.EmailAuthRequestDto;
 import jpabook.dashdine.dto.request.user.EmailRequestDto;
-import jpabook.dashdine.redis.RedisUtil;
 import jpabook.dashdine.service.user.query.UserQueryService;
+import jpabook.dashdine.util.RedisUtil;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
